requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dnspython==2.4.2
docutils==0.20.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
frozenlist==1.4.0
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
packaging==23.2
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinxawesome-theme==5.0.1
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stripe==7.5.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
websocket-client==1.6.4
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
tabulate==0.9.0